ENH: Also just register as a hypha service
